requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
pydantic==2.5.0
python-multipart==0.0.6